move: styles 폴더 이동 (#2)
