+++ src/styles/storeListStyles.css.ts
